separação de código para organização
